global: Change string position/length from unsigned int to size_t
Mainly to avoid truncating >4GB strings, which might potentially cause
some security holes. Normally there are other limits, which prevent such
excessive strings from being created in the first place.

I'm sure this didn't find everything. Maybe everything could be found with
compiler warnings. -Wconversion kind of does it, but it gives way too many
unnecessary warnings.

These were mainly found with:

grep " = strlen"
egrep "unsigned int.*(size|len)"

+++ src/auth/auth-request.c
@@ -1685,7 +1685,7 @@ void auth_request_set_field(struct auth_request *request,
 			    const char *name, const char *value,
 			    const char *default_scheme)
 {
-	unsigned int name_len = strlen(name);
+	size_t name_len = strlen(name);
 
 	i_assert(*name != '\0');
 	i_assert(value != NULL);
@@ -1895,7 +1895,7 @@ auth_request_userdb_import(struct auth_request *request, const char *args)
 void auth_request_set_userdb_field(struct auth_request *request,
 				   const char *name, const char *value)
 {
-	unsigned int name_len = strlen(name);
+	size_t name_len = strlen(name);
 	uid_t uid;
 	gid_t gid;
