+++ src/lib-mail/test-istream-qp-decoder.c
@@ -25,7 +25,7 @@ static void
 decode_test(const char *qp_input, const char *output, bool broken_input,
 	    unsigned int buffer_size)
 {
-	unsigned int qp_input_len = strlen(qp_input);
+	size_t qp_input_len = strlen(qp_input);
 	struct istream *input_data, *input;
 	const unsigned char *data;
 	size_t i, size;
