+++ src/lib/printf-format-fix.h
@@ -5,7 +5,7 @@
    used. If the format string was modified, it's returned from data stack. */
 const char *printf_format_fix(const char *format) ATTR_FORMAT_ARG(1);
 /* Like printf_format_fix(), except return also the format string's length. */
-const char *printf_format_fix_get_len(const char *format, unsigned int *len_r)
+const char *printf_format_fix_get_len(const char *format, size_t *len_r)
 	ATTR_FORMAT_ARG(1);
 /* Like printf_format_fix(), except the format string is written to data
    stack without actually allocating it. Data stack must not be used until
